LPS-63538 Move ServiceWrapperProxyUtil to the service package
