+++ portal-impl/src/com/liferay/portal/upgrade/v6_1_0/UpgradePermission.java
@@ -42,7 +42,7 @@ import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LoggingTimer;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.upgrade.ServiceWrapperProxyUtil;
+import com.liferay.portal.service.ServiceWrapperProxyUtil;
 
 import java.io.Closeable;
